+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.2.8]
+- Fix bug where "lock session" button does not show up on the first run
+
+## [2.2.7]
 - Add logging to file in user roaming appdata directory
 - Add ability to set custom text for Terms of Service on the message and in the "Show Details" section
+
+## [2.2.6]
 - Add ability to show time in system try icon and/or a "splash" in the corner of the screen
 
 ## [2.2.5]